+++ CHANGELOG.md
@@ -5,9 +5,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## [6.1.27] - 2024-12-16
+### Fixed
+- Cleaned up Image & Audio Result data dispatch [#4402](https://github.com/rokwire/illinois-app/issues/4402).
+- Fixed name pronunciation processing [#4402](https://github.com/rokwire/illinois-app/issues/4402).
+### Added
+- New Azure Assistant provider [#4519](https://github.com/rokwire/illinois-app/issues/4519).
+### Changed
+- Standardized return type of loadUserPhoto and loadUserNamePronunciation [#4402](https://github.com/rokwire/illinois-app/issues/4402).
+- Various fixes and improvements in Directory Profiles [#4402](https://github.com/rokwire/illinois-app/issues/4402).
+- GroupDetailPanel UI updates (in progress) [#4483](https://github.com/rokwire/illinois-app/issues/4483).
+
+## [6.1.26] - 2024-12-13
 ### Changed
 - Control wallet items visibility from app config settings [#4509](https://github.com/rokwire/illinois-app/issues/4509).
-
+- Sync photo / pronunciation URLs to user profile [#4402](https://github.com/rokwire/illinois-app/issues/4402).
+- Directory texts fixes/cleanup [#4402](https://github.com/rokwire/illinois-app/issues/4402).
+### Fixed
+- Fixed "view-dark" image resource.
+ 
 ## [6.1.25] - 2024-12-09
 ### Added
 - Added Directory section in Browse [#4402](https://github.com/rokwire/illinois-app/issues/4402).